--- spring-cloud-alibaba-examples/appactive-example/common/src/main/java/io/appactive/demo/common/Constants.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2013-2021 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.appactive.demo.common;
-
-/**
- * @author ChengPu raozihao
- * @description
- * @date 2022/8/15
- */
-public final class Constants {
-
-	/**
-	 * center flag.
-	 */
-	public static final String CENTER_FLAG = "center";
-
-	private Constants() {
-	}
-
-}
