+++ spring-cloud-alibaba-examples/appactive-example/product/src/main/java/com/alibaba/cloud/example/product/service/ProductServiceNormalImpl.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2013-2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alibaba.cloud.example.product.service;
+
+import java.util.List;
+
+import com.alibaba.cloud.example.common.entity.Product;
+import com.alibaba.cloud.example.common.entity.ResultHolder;
+import com.alibaba.cloud.example.common.service.ProductServiceNormal;
+import com.alibaba.cloud.example.product.repository.ProductRepository;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ProductServiceNormalImpl implements ProductServiceNormal {
+
+	@Autowired
+	ProductRepository productRepository;
+
+	@Value("${appactive.unit}")
+	private String unit;
+
+	@Override
+	public ResultHolder<List<Product>> list() {
+		return new ResultHolder<>(productRepository.findAll());
+	}
+
+}
