+++ spring-cloud-alibaba-examples/appactive-example/product/src/main/resources/bootstrap.properties
@@ -1,0 +1,9 @@
+spring.application.name=product
+spring.cloud.nacos.config.server-addr=127.0.0.1:8848
+
+#nacos certification information
+spring.cloud.nacos.username=nacos
+spring.cloud.nacos.password=nacos
+
+spring.cloud.nacos.config.refresh-enabled=true
+
