+++ spring-cloud-alibaba-examples/appactive-example/storage/src/main/resources/application.properties
@@ -1,20 +1,16 @@
-server.port=8087
-spring.application.name=product
+server.port=8085
+spring.application.name=storage
 spring.datasource.url=jdbc:mysql://127.0.0.1:3306/product?characterEncoding=utf8&useSSL=false&serverTimezone=GMT
 spring.datasource.username=root
 spring.datasource.password=root
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+#spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+spring.datasource.driver-class-name=io.appactive.db.mysql.driver.Driver
 spring.jpa.database-platform=org.hibernate.dialect.MySQL5InnoDBDialect
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=false
 appactive.unit=center
-appactive.app=product
+appactive.app=storage
 spring.cloud.nacos.discovery.server-addr=127.0.0.1:8848
-feign.client.config.default.connectTimeout=5000
-feign.client.config.default.readTimeout=5000
-spring.cloud.appactive.filter.unit-path=/detailHidden/*,/detail/*
 spring.cloud.appactive.filter.center-path=/buy/*
 spring.cloud.appactive.filter.normal-path=/*
-
-storage.ribbon.NFLoadBalancerRuleClassName =com.alibaba.cloud.appactive.consumer.AppactiveRule
-
