+++ spring-cloud-alibaba-examples/pom.xml
@@ -26,6 +26,7 @@
         <module>appactive-example/common</module>
         <module>appactive-example/frontend</module>
         <module>appactive-example/product</module>
+        <module>appactive-example/storage</module>
         <module>nacos-example/nacos-discovery-example</module>
         <module>nacos-example/nacos-config-example</module>
         <module>nacos-example/nacos-gateway-example</module>
